backend/requirements.txt
fastapi
uvicorn
PyPDF2
python-dotenv
langchain_groq
